+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.10.4dev
 * [Fix] Fix bug causing empty result on SQL with trailing semicolon and comment (#907)
+* [Fix] Fix bug %sql not parsing JSON arrow operators correctly (#918)
 
 * [Fix] Fixed bug that returns empty results when exception is raised from DB driver